requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0
google-generativeai==0.3.2
python-dotenv==1.0.1
replicate==0.22.0
click>=8.1.7
Pillow>=10.2.0 